Delete binary file
